Switch Api to v3.
