+++ api/GetConfig/sample.dat
@@ -1,0 +1,3 @@
+{
+    "name": "Azure"
+}
