--- api/local.settings.sample.json
@@ -1,13 +1,0 @@
-{
-  "IsEncrypted": false,
-  "Values": {
-    "AzureWebJobsStorage": "",
-    "FUNCTIONS_WORKER_RUNTIME": "node",
-    "AzureWebJobsFeatureFlags": "EnableWorkerIndexing",
-    "EnvName": "CDEV - JSON",
-    "IsFirebase": false,
-    "FireBaseAPIKey": "AIzaSyCubx8YGktd3hNv6_8wKGdRYaCqsmN3JU4",
-    "IsRestAPI": true,
-    "RESTLoggingUrl": "https://ngangularpracticesapi.azurewebsites.net/logger"
-  }
-}
