+++ api/package.json
@@ -6,9 +6,6 @@
     "start": "func start",
     "test": "echo \"No tests yet...\""
   },
-  "dependencies": {
-    "@azure/functions": "^4.0.0"
-  },
-  "devDependencies": {},
-  "main": "src/functions/*.js"
+  "dependencies": {},
+  "devDependencies": {}
 }
